rename '_utilities' to '_utils'
